% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeastJar.R
\docType{package}
\name{BeastJar-package}
\alias{BeastJar-package}
\title{BeastJar: JAR Dependency for MCMC Using 'BEAST'}
\description{
Provides JAR to perform Markov chain Monte Carlo (MCMC) inference using the popular Bayesian Evolutionary Analysis by Sampling Trees 'BEAST X' software library of Baele et al (2025) \doi{10.1038/s41592-025-02751-x}. 'BEAST X' supports auto-tuning Metropolis-Hastings, slice, Hamiltonian Monte Carlo and Sequential Monte Carlo sampling for a large variety of composable standard and phylogenetic statistical models using high performance computing. By placing the 'BEAST X' JAR in this package, we offer an efficient distribution system for 'BEAST X' use by other R packages using CRAN.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/beast-dev/BeastJar}
}

}
\author{
\strong{Maintainer}: Marc A. Suchard \email{msuchard@ucla.edu} [copyright holder]

Other contributors:
\itemize{
  \item Andrew Rambaut [copyright holder]
  \item Alexei J. Drummond [copyright holder]
}

}
\keyword{internal}
