% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeastJar.R
\docType{package}
\name{BeastJar}
\alias{BeastJar}
\title{BeastJar}
\description{
Convenient packaging of the Bayesian Evolutionary Analysis Sampling Trees (BEAST) software package
to facilitate Markov chain Monte Carlo sampling techniques including Hamiltonian Monte Carlo,
bouncy particle sampling and zig-zag sampling.
}
