% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeastMcmc.R
\name{simulateNormalDistribution}
\alias{simulateNormalDistribution}
\title{Example MCMC simulation using BEAST}
\usage{
simulateNormalDistribution(mean = 0, sd = 1, chainLength = 10000,
  subSampleFrequency = 10)
}
\arguments{
\item{mean}{Mean of target normal distribution}

\item{sd}{Standard deviation of target normal distribution}

\item{chainLength}{Markov chain length to simulate}

\item{subSampleFrequency}{Frequency at which to retain samples from the Markov chain}
}
\description{
This function generates a Markov chain to sample from a simple normal distribution.
It uses a random walk Metropolis kernel that is auto-tuning.
}
